@@ -5,7 +5,7 @@ description = Metric as a Service
 long_description = file: README.rst
 long_description_content_type = text/x-rst
 author = Gnocchi developers
-python_requires = >=3.6
+python_requires = >=3.8
 classifier =
     Intended Audience :: Information Technology
     Intended Audience :: System Administrators
@@ -14,9 +14,9 @@ classifier =
     Programming Language :: Python
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.11
     Topic :: System :: Monitoring
 
 [options]
@@ -36,7 +36,6 @@ install_requires =
     futures; python_version < '3'
     jsonpatch
     cotyledon>=1.5.0
-    six
     stevedore
     ujson
     voluptuous>=0.8.10
@@ -47,7 +46,9 @@ install_requires =
     Paste
     PasteDeploy
     daiquiri
-    pyparsing>=2.2.0
+    # TODO(tobias-urdin): Evaluate unpin later
+    # https://github.com/pyparsing/pyparsing/issues/504
+    pyparsing>=2.2.0,<3.1.0
     lz4>=0.9.0
     tooz>=1.38
     cachetools
@@ -85,7 +86,6 @@ prometheus =
 amqp1 =
     python-qpid-proton>=0.17.0
 doc =
-    chardet<4
     sphinx
     sphinx_rtd_theme
     sphinxcontrib-httpdomain
@@ -102,7 +102,7 @@ test =
     testscenarios
     testresources>=0.2.4 # Apache-2.0/BSD
     testtools>=0.9.38
-    WebTest>=2.0.16
+    WebTest>=3.0
     keystonemiddleware>=4.0.0,!=4.19.0
     wsgi_intercept>=1.4.1
     xattr!=0.9.4  # https://github.com/gnocchixyz/gnocchi/issues/951
