@@ -19,7 +19,7 @@ import logging
 import daiquiri
 from oslo_config import cfg
 from oslo_db import options as db_options
-from six.moves.urllib import parse as urlparse
+from urllib import parse as urlparse
 
 import gnocchi
 from gnocchi import archive_policy
